Add Time API.
We need some automatic testing infrastructure for these APIs. For now, the
test just prints the output of some corner cases, but these should be tested
automatically eventually.

There are some things, namely the list of timezones and date formats, that
are not properly implemented.  These require synchronous calls which are not
currently supported by the message passing extension infrastructure.