notification: make NotificationManager use string ids
Instead of exposing priv_id to the NotificationContext, let's use the
ids that it gets from JavaScript. All the mapping happens in the same
place and we keep only one datastructure.

This have the side effect of freeing us from a race condition when
running the client callback (runs on the main extension thread instead
of the context thread). This solution not always scale, in the long
run we probably want to expose some simplified version of MessageLoops.