Various fixes in the power API.
- Disable for now the tests which are turning off the screen. It is impossible
afterwards to get the results as after unlocking the screen Xwalk is not visible.
We can enable these tests again when the integration with the VM is done.

- Fix state request API by using pm_change_state and pm_lock_state. I'm investigating
with another Intel fellow what exactly we should call.

- Workaround C API calls not working correctly. Any C API using vconf can't work
properly from the extensions thread. I believe vconf needs to be in the main thread.
We use a glib timeout callback as a workaround. This can be removed when the extensions
are running in a separate process.