[Bluetooth] Remove setPowered implementation.
Due to Tizen 2.x' platform-defined behavior, we are removing here
the adapter.setPowered API implementation. Modifying this property on
this platform actually makes it kill the bluetooth daemon and (soft) rfkill
the bluetooth module. After that we lose access even to the DBus interfaces
related to BlueZ, leading us to a inconsistent state.

In order to land a proper fix for this we need to first have
more control over our sync message return, so we can work-around
this platform issue and handle the bluetooth init/deinit during the
next tizen.bluetooth.getDefaultAdapter() (i.e. after a setPowered(false),
for instance).