bluetooth: Fix not adding a device to the known devices list
When the device is already known to BlueZ, an DeviceUpdated() message
will be sent, and it means that the device should be added directly
to the known_devices list that is kept.

One example of when this could happen is when there are already paired
devices when the extension starts.