bluetooth: Use the same DeviceMap as BlueZ 5
As we want to keep track of changes in org.bluez.Device's objects,
it is very convenient that we keep GDBusProxies for each device.

And the plan is to have the key of the map to be the object path of
the device, so when an signal is emitted it can be immediately
recognized which device has changed.