bluetooth: Keep track of org.bluez.Device objects
The process is more involved for BlueZ 4 than for BlueZ 5. Before we can
notify upper layers that we have a new device we need at least to have its
address, for BlueZ 4, this requires that we send a D-Bus message to get its
properties.

For this process we keep a map of object paths to addresses, so when a
change is detected on one object path, it can be detected and we can
send a message to the JS side.