bluetooth: Fix notifying the same device multiple times on discovery
With the changes to what information we keep on the C++ side of the
extension, there was a side effect that the same defice was reported
multiple times on the ondevicefound() Discovery callback.

The fix is to handle marking the device as "new" on the JS side, as it
already has the known_devices and found_devices list, so it can take the
correct decision.