demo and examples [re]packaging
This patch reorganizes the packaging for demo and example. With that
we now generate 3 packages: 1) the exentions itself, 2) examples and
3) demo.

The resulting packages are:
   tizen-extensions-crosswalk-${version}-${release}.${arch}.rpm
   tizen-extensions-crosswalk-bluetooth-demo-${version}-${release}.${arch}.rpm
   tizen-extensions-crosswalk-examples-${version}-${release}.${arch}.rpm

For both demo and examples packages a manifest xml file is generated
so we can have a launcher icon within tizen menu-screen.

Installation:
The extensions rpm installation remains the same(i.e rpm -Uvh --force
tizen-extensions-crosswalk-${version}-${release}.${arch}.rpm).

The additional packages(demo and examples) must be installed with
pkgcmd, like:
   $ pkgcmd -i -t rpm -p tizen-extensions-crosswalk-bluetooth-demo-0.2-0.i586.rpm -q
   $ pkgcmd -i -t rpm -p tizen-extensions-crosswalk-examples-0.2-0.i586.rpm -q