[Power] Fix a TCT failure with isScreenOn()
TCT tests are expecting turnScreenOn/Off to have a synchronous behavior
which in fact means that querying the screen state right after should
return the correct value. In fact the underlaying platform does not
necessarily switch on or off the screen right away, it may take few
ms. This patch add a synchronous behavior to isScreenOn by
making sure to return when the platform informed us about the state
change of the screen (if called previously with turnScreenOn/Off).
This is slightly better than a busy wait.