[Notification] Do not keep separated copy of notifications
Notification object is essentialy as a "property bag" to be used in
conjunction with tizen.notification.* functions. Now that we have
v8tools.forceSetProperty() we can keep certain properties read-only and
do not need to worry about user changing the notification, so no need to
keep our own copy of the data.

Note that since Notification API doesn't rely on GC to identify
notifications that should be removed, one needs to explicitly call the
remove function, or wait for user removal via the UI.