[Notification] Make notification IDs assigned by native side
We share the same id-space for notifications in Mobile, so makes more
sense reuse the assigned IDs from the native notification library. This
reduces the mapping work we have to do.

To achieve this we had to change the 'NotificationPost' message to be
synchronous, enabling us to throw errors correctly when there was a
failure posting the notification.