[Notification] Add update() support for Desktop
Added a new message 'NotificationUpdate' that will update an existing
notification. This can be tested in the example by clicking the
"Update (N)" button.

This commit creates a helper struct to ease extracting the parameters
from the message we receive. As we start to pass more parameters, we
fill this struct.

Note that Desktop will support only a subset of all possible properties
that a notification can have.