bluetooth: Simplify PropertyChanged handling
There's no need to wait for the 'PropertyChanged' signal before
sending the reply to the caller. BlueZ guarantees that when
'SetProperty' returns the property will be applied on the adapter (or
an error will be returned).

The 'PropertyChanged' should be used for monitoring if another
application has changed the adapter's state.