Enable -std=c++0x
Tizen 2.x compiler is g++ 4.5.3 which already support some of the C++11
features like 'auto' and lambdas. It is fine to use those features in
the extensions.

Unfortunately there's an issue when compiling code using bluetooth.h in
C++11, so it's it disable for the tizen_bluetooth extension for now.