[Bookmark] Implement JS API on top of a fake provider
This patch adds a complete JS API, including all the error handling
covered by TCT tests. The public API is implemented in terms of a
BookmarkProvider object.

At the moment there's a FakeBookmarkProvider implemented, that stores
all the data in JavaScript. This was used to bootstrap the JS API and
implement the error handling. As the name implies, it doesn't actually
store the bookmarks in the real database.

Next step is to replace 'FakeBookmarkProvider' with a real provider
object that communicates with the native side of the extension.