Add XW_Extension_EntryPoints.h
This defines an internal interface in Crosswalk to setup "entry points"
of the API, so that Crosswalk can load the JS API on demand.

For example, 'tizen.time' will be loaded when the web content tries to
access 'tizen.time' in JS, but also when the web content tries to access
'tizen.TZDate'. This extra symbols are the entry points.