Add entry point information for extensions using old adapter
Entry point information is used by Crosswalk to know what symbols
can trigger the load of the extension, in case it decides to load on
demand. The extension name is implicitly an entry point, so we just
set the extra entry points.

This should cause no changes if the Crosswalk doesn't support the
interface for setting entry points yet -- like Crosswalk 1.