[Messageport] Add Messageport implementation.
Tests are passing, but since application installation support isn't ready,
it requires a few hacks for this to actually work.

This includes writing a short wrapper program that correctly passes the path
to Tizen extensions, adding a mocked version of
`tizen.application.getCurrentApplication()` that returns the standard
`tct-messageport` package ID, installing this package, and modifying
`channel-service` so that the application ID is hardcoded.