[Tizen] Fix trying to define properties on a undefined object
When the extension is loaded, the object 'tizen' where the extension
code is referring to may not exist yet, it is only guaranteed to exist
after the extension is loaded. So instead of referring to the 'tizen'
object we need to refer to it as 'exports' that will be translated to
'tizen' after the extension is loaded.