Add needed SMACK rules.
This commit install a smack rules for t-e-c "by hand",
instead of using the manifest file. We do it this way because
t-e-c is a "special" case: it is not a web app, thus no real
app id available. It will fallback to the case in which its
ExtensionProcess will run with the default SMACK label
"crosswalk::ep".

This commit makes all extensions work again but Bluetooth,
Download and Network Bearer management, all of which are affected
by the DBus System Bus label bug (please refer to https://github.com/jeez/dbus-bluez4-test).