Build desktop, Tizen mobile and IVI profiles properly.
Replace TIZEN_MOBILE definitions for TIZEN, change spec file to test on the %profile
variable for ivi and mobile (defined in build.xml and propagated to build.conf)
instead of _with_ivi. In the gyp files, 'extension_host_os' tests have been replaced
with testing 'tizen'. In source code, TIZEN_MOBILE used to be assumed to work with
 IVI, considering mobile equal to IVI, which is not the case with every extension.
With this change, 'extension_host_os' is set to 'desktop' by default.
If a tizen repository is present, TIZEN will be defined to 1. In addition, TIZEN_IVI
is defined for IVI and TIZEN_MOBILE is defined for mobile. All extensions which
should work on all Tizen profiles, should use TIZEN, others should use the more
specific TIZEN_MOBILE and TIZEN_IVI.