Upgrade after crosswalk always use tizen_app_id on tizen, fix cannot get current app info and dbus proxy.
PR#2261 : https://github.com/crosswalk-project/crosswalk/pull/2261

After PR#2261, we always get the app_id which will equal to the
app_id storaged in pkgmgr DB, so there is no need to convert
pkg_id to app_id.

Because object path with '.' is an invalid object path, so
after PR#2261 the object path of running application on dbus
was changed so we need to change it too.

BUG=XWALK-1925